Basic simple of html password
This is the basic password example. So I've been using a different input type for the password and the username. as you see the code. And the fieldset would work really well which is like a little vbox and also you can change or ask me questions if you want!!